Migration to Styled-Components
