+++ package.json
@@ -58,6 +58,7 @@
     "@types/react-test-renderer": "16.8.1",
     "@types/react-transition-group": "2.0.16",
     "@types/storybook__react": "4.0.1",
+    "@types/styled-components": "4.1.8",
     "@types/styled-jsx": "2.2.8",
     "babel-jest": "24.8.0",
     "babel-plugin-module-resolver": "3.2.0",
@@ -92,6 +93,7 @@
     "lodash.isstring": "4.0.1",
     "react-day-picker": "7.3.0",
     "react-transition-group": "2.6.1",
+    "styled-components": "4.3.2",
     "styled-jsx": "3.2.1"
   },
   "peerDependencies": {
