+++ src/index.tsx
@@ -1,105 +1,53 @@
-import AutoComplete from 'autoComplete'
-import Avatar from 'avatar'
-import Badge from 'badge'
-import Button from 'button'
-import ButtonGroup from 'buttonGroup'
-import Caption from 'caption'
-import ConfirmationModal from 'confirmationModal'
-import DatePicker from 'datePicker'
-import Drawer from 'drawer'
-import DropdownButton from 'dropdownButton'
-import EmptyState from 'emptyState'
-import HamburgerButton from 'hamburgerButton'
-import ItemAction from 'itemAction'
-import ItemCheckbox from 'itemCheckbox'
-import ItemChoice from 'itemChoice'
-import ItemData from 'itemData'
-import ItemInfo from 'itemInfo'
-import ItemRadio from 'itemRadio'
-import ItemRadioGroup from 'itemRadioGroup'
-import ItemsList from 'itemsList'
-import Itinerary from 'itinerary'
-import Loader from 'loader'
-import Menu from 'menu'
-import Message from 'message'
-import MessagingSummaryItem from 'messagingSummaryItem'
-import Modal from 'modal'
-import PhoneField from 'phoneField'
-import Profile from 'profile'
-import Proximity from 'proximity'
-import PushInfo from 'pushInfo'
-import Rating from 'rating'
-import SelectField from 'selectField'
-import Snackbar from 'snackbar'
-import Stars from 'stars'
-import Stepper from 'stepper'
-import SuccessModal from 'successModal'
-import Tabs from 'tabs'
-import Text from 'text'
-import TextField from 'textField'
-import TimePicker from 'timePicker'
-import Title from 'title'
-import ToggleButton from 'toggleButton'
-import TopBar from 'topBar'
-import Transitions from 'transitions'
-import TripCard from 'tripCard'
-import UneditableTextField from 'uneditableTextField'
-import Why from 'why'
+export { default as AutoComplete } from 'autoComplete'
+export { default as Avatar } from 'avatar'
+export { default as Badge } from 'badge'
+export { default as Button } from 'button'
+export { default as ButtonGroup } from 'buttonGroup'
+export { default as Caption } from 'caption'
+export { default as ConfirmationModal } from 'confirmationModal'
+export { default as DatePicker } from 'datePicker'
+export { default as Drawer } from 'drawer'
+export { default as DropdownButton } from 'dropdownButton'
+export { default as EmptyState } from 'emptyState'
+export { default as HamburgerButton } from 'hamburgerButton'
+export { default as ItemAction } from 'itemAction'
+export { default as ItemCheckbox } from 'itemCheckbox'
+export { default as ItemChoice } from 'itemChoice'
+export { default as ItemData } from 'itemData'
+export { default as ItemInfo } from 'itemInfo'
+export { default as ItemRadio } from 'itemRadio'
+export { default as ItemRadioGroup } from 'itemRadioGroup'
+export { default as ItemsList } from 'itemsList'
+export { default as Itinerary } from 'itinerary'
+export { default as Loader } from 'loader'
+export { default as Menu } from 'menu'
+export { default as Message } from 'message'
+export { default as MessagingSummaryItem } from 'messagingSummaryItem'
+export { default as Modal } from 'modal'
+export { default as PhoneField } from 'phoneField'
+export { default as Profile } from 'profile'
+export { default as Proximity } from 'proximity'
+export { default as PushInfo } from 'pushInfo'
+export { default as Rating } from 'rating'
+export { default as SelectField } from 'selectField'
+export { default as Snackbar } from 'snackbar'
+export { default as Stars } from 'stars'
+export { default as Stepper } from 'stepper'
+export { default as SuccessModal } from 'successModal'
+export { default as Tabs } from 'tabs'
+export { default as Text } from 'text'
+export { default as TextField } from 'textField'
+export { default as TimePicker } from 'timePicker'
+export { default as Title } from 'title'
+export { default as ToggleButton } from 'toggleButton'
+export { default as TopBar } from 'topBar'
+export { default as Transitions } from 'transitions'
+export { default as TripCard } from 'tripCard'
+export { default as UneditableTextField } from 'uneditableTextField'
+export { default as Why } from 'why'
 
-import branding from '_utils/branding'
-import flush, { flushToHTML } from 'styled-jsx/server'
+export { default as branding } from '_utils/branding'
 
 export * from 'icon/index'
 
-export {
-  AutoComplete,
-  Avatar,
-  Badge,
-  Button,
-  ButtonGroup,
-  Caption,
-  ConfirmationModal,
-  DatePicker,
-  Drawer,
-  DropdownButton,
-  EmptyState,
-  HamburgerButton,
-  ItemAction,
-  ItemCheckbox,
-  ItemChoice,
-  ItemData,
-  ItemInfo,
-  ItemRadio,
-  ItemRadioGroup,
-  ItemsList,
-  Itinerary,
-  Loader,
-  Menu,
-  Message,
-  MessagingSummaryItem,
-  Modal,
-  PhoneField,
-  Profile,
-  Proximity,
-  PushInfo,
-  Rating,
-  SelectField,
-  Snackbar,
-  Stars,
-  Stepper,
-  SuccessModal,
-  Tabs,
-  Text,
-  TextField,
-  TimePicker,
-  Title,
-  ToggleButton,
-  TopBar,
-  Transitions,
-  TripCard,
-  UneditableTextField,
-  Why,
-  branding,
-  flush,
-  flushToHTML,
-}
+export { default as flush, flushToHTML } from 'styled-jsx/server'
